Monkey-patch in `puts` and `print`
For simplicity, this patch make `puts` and `print` behave as expected.

Was a little confused with running `bin/sandbox` giving me an empty `stdout`
when I tried to run a simple hello world program.
```rb
puts "Hello, World!"
```
which gave result
```
<struct EnterpriseScriptService::Result output=nil, stdout="", stat=#<struct EnterpriseScriptService::Stat instructions=0, memory=0, bytes_in=0, time=0, execution_time_us=0, total_instructions=0>, measurements={:in=>404, :mem=>28, :init=>6163, :sandbox=>411, :decode=>1, :inject=>1, :lib=>0, :compile=>80}, errors=[#<EnterpriseScriptService::EngineIllegalSyscallError: engine terminated with signal 31>]>
```

Upon closer inspection, the `EngineIllegalSyscallError` is because we're trying
to print to stdout which is blocked by the `seccomp` sandbox. After a bit of
digging, the right way to do output is to use the instance variable
`@stdout_buffer` which we use in our monkey-patches.